+++ computenestsupplier-20210521/core/Models/CreateArtifactShrinkRequest.cs
@@ -10,6 +10,13 @@ namespace AlibabaCloud.SDK.ComputeNestSupplier20210521.Models
 {
     public class CreateArtifactShrinkRequest : TeaModel {
         /// <summary>
+        /// <para>The build properties of the artifact, utilized for hosting and building the deployment package.</para>
+        /// </summary>
+        [NameInMap("ArtifactBuildProperty")]
+        [Validation(Required=false)]
+        public string ArtifactBuildPropertyShrink { get; set; }
+
+        /// <summary>
         /// <para>The ID of the deployment package.</para>
         /// 
         /// <b>Example:</b>