+++ computenestsupplier-20210521/core/Models/RejectServiceUsageResponseBody.cs
@@ -10,6 +10,8 @@ namespace AlibabaCloud.SDK.ComputeNestSupplier20210521.Models
 {
     public class RejectServiceUsageResponseBody : TeaModel {
         /// <summary>
+        /// <para>The request ID.</para>
+        /// 
         /// <b>Example:</b>
         /// <para>4DB0F536-B3BE-4F0D-BD29-E83FB56D550C</para>
         /// </summary>